whalebot/gbot/leviathan_py/requirements.txt
numpy
pybit
google-generativeai
python-dotenv
aiohttp
websockets
pytest
pytest-asyncio
pytest-cov
pytest-xdist
radon
pandas
